A few minor but useful changes:
Reprog/SD_reprog.c
If no SN at startup, check for the system/*.srn file and use that. 
  - This is particularly helpful after reflashing firmware from probe
    when SN gets wiped out.  This means that updateSN() in SD_reprog.c
    now takes a path parameter.

Reprog/SD_reprog.c
Bug Fix: updateSN() now properly handles filename '-erase-.srn' so it
    erases the SN, setting it to 0s. Same behavior as using probe to 
    reflash firmware.

device.c, device.h, testPCB.c and startup.c
writeVersionToDisk() also now takes a path parameter so that we can 
    use SYSTEM_PATH when we have it (from config) and 
    DEFAULT_SYSTEM_PATH when we don't.

device.c
Fixed bug in getDeviceSN(), which was returning a pointer without 
    checking if the string terminated any time soon.  This was a 
    problem when SN was a bunch of 0xFFFFs.  Now erase uses 0x0000,
    so this bug fix isn't redundant protection for now.

device.c
confirmSNonDisk() now wipes out any old *.srn files that don't match
    the NOR flash SN.  This is especially important now that startup()
    will set the SN if it is absent from the first *.srn file it finds.

main.c, startup.c
Moved LOG_FILE = DEFAULT_LOG_FILE into setDefaults(),
    and moved setDefaults() into main.c. Logging can still be turned 
    off with a config entry that has no path (e.g. LOG_FILE:)

startup.c
Bug Fix: startup() now sets inspect = 0 at top of function. This is 
    necessary 

startup.c
Turn on red & green LEDs immediately upon startup().  Users will get 
    a faster indication that the device is on, and cold start will 
    briefly blink the lights before halting.


startup.c
Extra log messages: probe firmware update, 
                    use of disaster_config_strings,
                    use of backup config file,
                    when user tries to play an empty category

startup.c
Create 'inspect' filename when system.img is found to trigger any 
    other checks that might be needed after reprogamming takes place.

+++ device/software/device/trunk/firmware/Reprog/SD_reprog.c
@@ -159,15 +159,18 @@ ReprogFailed(flash *fp)
 }
 */
 
-void updateSN() {
+void updateSN(char* path) {
 	extern void write_app_flash(int *, int, unsigned int);
 	struct f_info file_info;
 	char *newSN;
 	char pathFrom[PATH_LENGTH], pathTo[PATH_LENGTH];
 	int ret, goodSN, sameSN;
-	
+	int reuseSystemSRNfile;
+		
 	// see if a .srn file is present
-	ret =_findfirst((LPSTR)UPDATE_FP SERIAL_FN, &file_info, D_FILE);
+	strcpy(pathFrom,path);
+	strcat(pathFrom,SERIAL_FN);
+	ret =_findfirst((LPSTR)pathFrom, &file_info, D_FILE);
 	if(ret >= 0) {
 		char *dot;
 		int lenEraseCode = strlen((char *)ERASE_SN_CODE);
@@ -182,18 +185,19 @@ void updateSN() {
 			newSN += lenEraseCode;
 		}
 	 	// check that sn begins with "srn."
-	 	goodSN = !strncmp(newSN,CONST_TB_SERIAL_PREFIX,CONST_TB_SERIAL_PREFIX_LEN);
-	 	
-		// move srn. file to system directory to allow file-based inspection of device SN
-		strcpy(pathFrom,(char *)UPDATE_FP);
-		strcat(pathFrom,file_info.f_name);
-		strcpy(pathTo,(char *)DEFAULT_SYSTEM_PATH);
-		ret = mkdir((LPSTR)pathTo);
-		strcat(pathTo,newSN + CONST_TB_SERIAL_PREFIX_LEN);
-		ret = rename((LPSTR)pathFrom,(LPSTR)pathTo);
-		if (ret == -1) {
-			unlink((LPSTR)pathFrom);
-		}
+	 	if (strcmp(path,DEFAULT_SYSTEM_PATH)) {
+		 	goodSN = !strncmp(newSN,CONST_TB_SERIAL_PREFIX,CONST_TB_SERIAL_PREFIX_LEN);
+			// move srn. file to system directory to allow file-based inspection of device SN
+			strcpy(pathFrom,(char *)UPDATE_FP);
+			strcat(pathFrom,file_info.f_name);
+			strcpy(pathTo,(char *)DEFAULT_SYSTEM_PATH);
+			ret = mkdir((LPSTR)pathTo);
+			strcat(pathTo,newSN + CONST_TB_SERIAL_PREFIX_LEN);
+			ret = rename((LPSTR)pathFrom,(LPSTR)pathTo);
+			if (ret == -1) 
+				unlink((LPSTR)pathFrom);
+	 	} else
+			reuseSystemSRNfile = goodSN = 1; // "srn." won't be at the beginning of the file in /system	
 
 	 	//check that new sn is not the same as old - force goodSN to 0 if it is the same
 		if (*(char *)TB_SERIAL_NUMBER_ADDR) // otherwise 0-length string will look equal
@@ -217,11 +221,22 @@ void updateSN() {
 					vlen = 0;
 				else
 					*dot = 0;
-				write_app_flash((int *)newSN, vlen, (unsigned int)0xffff);
+				if (reuseSystemSRNfile) {
+					// if reusing the /system *.srn file, then we must manually add the srn prefix
+					strcpy(pathFrom,CONST_TB_SERIAL_PREFIX);
+					strcat(pathFrom,newSN);
+					vlen = strlen(pathFrom);
+
+					write_app_flash((int *)pathFrom, vlen, (unsigned int)0x0000);
+				} else
+					write_app_flash((int *)newSN, vlen, (unsigned int)0x0000);
 			}
 			if (flagErase) {
 				unlink((LPSTR)pathTo);
 			}
+	 	} else if (flagErase) {
+	 		// no SN - just erase
+	 		write_app_flash((int *)"", 0, (unsigned int)0x0000);
 	 	}
 	}
 }
